+++ Actions/DownloadProjectDependencies/DownloadProjectDependencies.Action.ps1
@@ -38,6 +38,7 @@ function DownloadDependenciesFromCurrentBuild {
 
     Write-Host "Downloading dependencies for project '$project'"
 
+    $projectDependencies | ConvertTo-Json | Out-Host
     $dependencyProjects = @()
     if ($projectDependencies.Keys -contains $project) {
         $dependencyProjects = @($projectDependencies."$project")
@@ -46,7 +47,7 @@ function DownloadDependenciesFromCurrentBuild {
     Write-Host "Dependency projects: $($dependencyProjects -join ', ')"
 
     # For each dependency project, calculate the corresponding probing path
-    $dependeciesProbingPaths = @()
+    $dependenciesProbingPaths = @()
     foreach($dependencyProject in $dependencyProjects) {
         Write-Host "Reading settings for project '$dependencyProject'"
         $dependencyProjectSettings = ReadSettings -baseFolder $baseFolder -project $dependencyProject
@@ -70,7 +71,7 @@ function DownloadDependenciesFromCurrentBuild {
             $baseBranch = $ENV:GITHUB_REF_NAME
         }
 
-        $dependeciesProbingPaths += @(@{
+        $dependenciesProbingPaths += @(@{
             "release_status"  = "thisBuild"
             "version"         = "latest"
             "buildMode"       = $dependencyBuildMode
@@ -85,7 +86,7 @@ function DownloadDependenciesFromCurrentBuild {
 
     # For each probing path, download the dependencies
     $downloadedDependencies = @()
-    foreach($probingPath in $dependeciesProbingPaths) {
+    foreach($probingPath in $dependenciesProbingPaths) {
         $buildMode = $probingPath.buildMode
         $project = $probingPath.projects
         $branch = $probingPath.branch